pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "audio-recordings-manager"
version = "0.1.0"
description = "Self-hosted audio recording manager with automatic transcription"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "watchdog",
    "faster-whisper",
    "sqlalchemy",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "mypy",
    "ruff",
    "types-PyYAML",
    "pre-commit",
    "playwright",
    "pytest-playwright",
]

[tool.hatch.build.targets.wheel]
packages = ["mnemovox"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
